Arreglado el many to many

+++ Semana15-Delete/src/main/java/org/accesodatos/semana15Delete/dtos/response/AsignaturaCalificacionDTO.java
@@ -1,9 +1,10 @@
 package org.accesodatos.semana15Delete.dtos.response;
 
 import lombok.Data;
+import java.math.BigDecimal;
 
 @Data
 public class AsignaturaCalificacionDTO {
     private String asignatura;
-    private Double calificacion;
+    private BigDecimal calificacion;
 }
